Separated out model factory instance into separate class, added more documentation.
